requirements.txt
Flask==2.3.3
Pillow==9.5.0
gunicorn==21.2.0